Add files via upload
In JavaScript, the map() method is used to create a new array by applying a function to each element of the original array. It doesn't change the original array, but instead, it creates and returns a new array with the results of calling the provided function on every element in the original array.